+++ packages/lib/src/index.ts
@@ -2,3 +2,4 @@ export const isoNow = () => new Date().toISOString();
 export * from "./flags.data";
 export * from "./random.util";
 export * from "./history.store";
+export * from "./flags.pool";